adding basic post voting

+++ routers/post.py
@@ -3,7 +3,7 @@ from time import sleep
 from typing import Optional, List
 
 from fastapi import Response, status, HTTPException, Depends, APIRouter
-
+from sqlalchemy import func
 from sqlalchemy.orm import Session
 
 import models.logging_config
@@ -11,7 +11,7 @@ from models import db_models
 from models import db_schemas
 from app import oauth2
 from app.database import get_db
-from app.utils import *
+from app.utils.utils import *
 
 
 ###############################  SETUP  #################################
@@ -23,18 +23,49 @@ router = APIRouter(prefix="/posts", tags=["Posts"])
 
 
 # GET ALL POSTS
-@router.get("/", status_code=status.HTTP_200_OK, response_model=List[db_schemas.Post])
-def get_posts(db: Session = Depends(get_db)):
+@router.get("/", status_code=status.HTTP_200_OK, response_model=List[db_schemas.PostOut])
+def get_posts(
+    db: Session = Depends(get_db),
+    current_user: db_models.User = Depends(oauth2.get_current_user),
+    limit: int = 10,
+    skip: int = 0,
+    search: Optional[str] = "",
+):
 
-    posts = db.query(db_models.Post).all()
-    return posts
+    # posts = db.query(db_models.Post).filter(db_models.Post.owner_id == current_user.id).all()
+
+    results = (
+        db.query(
+            db_models.Post, func.count(db_models.Vote.post_id).label("votes_count")
+        )
+        .join(db_models.Vote, db_models.Vote.post_id == db_models.Post.id, isouter=True)
+        .group_by(db_models.Post.id)
+        .filter(db_models.Post.title.contains(search))
+        .limit(limit)
+        .offset(skip)
+        .all()
+    )
+
+    return results
 
 
 # GET ONE POST
-@router.get("/{id}", status_code=status.HTTP_200_OK, response_model=db_schemas.Post)
-def get_post(id: int, db: Session = Depends(get_db)):
+@router.get("/{id}", status_code=status.HTTP_200_OK, response_model=db_schemas.PostOut)
+def get_post(
+    id: int,
+    db: Session = Depends(get_db),
+    current_user: db_models.User = Depends(oauth2.get_current_user),
+):
 
-    post = db.query(db_models.Post).filter(db_models.Post.id == id).first()
+    post = (
+        db.query(
+            db_models.Post, func.count(db_models.Vote.post_id).label("votes_count")
+        )
+        .join(db_models.Vote, db_models.Vote.post_id == db_models.Post.id, isouter=True)
+        .group_by(db_models.Post.id)
+        .filter(db_models.Post.id == id)
+        .first()
+    )
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
     return post
@@ -45,10 +76,10 @@ def get_post(id: int, db: Session = Depends(get_db)):
 def create_post(
     post: db_schemas.PostCreate,
     db: Session = Depends(get_db),
-    user_id: int = Depends(oauth2.get_current_user),
+    current_user: db_models.User = Depends(oauth2.get_current_user),
 ):
-
-    new_post = db_models.Post(**post.model_dump())
+    print(f"printing current_user: {current_user.email}")
+    new_post = db_models.Post(**post.model_dump(), owner_id=current_user.id)
     db.add(new_post)
     db.commit()
     db.refresh(new_post)
@@ -57,12 +88,23 @@ def create_post(
 
 # DELETE ONE POST
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_post(id: int, db: Session = Depends(get_db)):
+def delete_post(
+    id: int,
+    db: Session = Depends(get_db),
+    current_user: db_models.User = Depends(oauth2.get_current_user),
+):
 
     post_query = db.query(db_models.Post).filter(db_models.Post.id == id)
     post = post_query.first()
+
     if post == None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
+    if post.owner_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="not authorized to delete a post",
+        )
+
     post_query.delete(synchronize_session=False)
     db.commit()
     return Response(status_code=status.HTTP_204_NO_CONTENT)
@@ -72,13 +114,24 @@ def delete_post(id: int, db: Session = Depends(get_db)):
 @router.put(
     "/{id}", status_code=status.HTTP_202_ACCEPTED, response_model=db_schemas.Post
 )
-def update_post(id: int, post: db_schemas.PostCreate, db: Session = Depends(get_db)):
+def update_post(
+    id: int,
+    updated_post: db_schemas.PostCreate,
+    db: Session = Depends(get_db),
+    current_user: db_models.User = Depends(oauth2.get_current_user),
+):
 
     post_query = db.query(db_models.Post).filter(db_models.Post.id == id)
+    post = post_query.first()
 
-    if not post_query.first():
+    if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
+    if post.owner_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="not authorized to make changes to a post",
+        )
 
-    post_query.update(post.model_dump(), synchronize_session=False)
+    post_query.update(updated_post.model_dump(), synchronize_session=False)
     db.commit()
     return post_query.first()
